test(assets/unzip): update test code to be independent of the `unzipper` package
Previous tests have mocked up the `Extract` function of the `unzipper` package.
This test does not confirm that the same functionality is maintained by updating the `src/assets/unzip.js` file.
It also does not test whether the previous code and the new code extracted the ZIP file correctly.

So I rewrote the test code.
By mocking the file system and the HTTP request, we can now do the same test no matter which unzip library we use.

Added dependencies:
+ mock-fs
+ nock

The rewritten test code was tested with the old `src/assets/unzip.js` before the update.
If necessary, check out this commit using the `git checkout` command and test the old `src/assets/unzip.js` with the updated test code.