test(assets/unzip): reduce test code differences
The test code before my changes used Jest mock functions.
However, I rewrote it into very unreadable test code using the `Promise` constructor.

This is because initially I could not think of a way to write an asynchronous test that waited until the mock function was called.
But now I have found a way.

So I will modify the test code to make the differences smaller.